refactor : 모듈, 디렉토리 명 수정
