+++ app/api/show-api/src/main/java/com/example/publish/message/TicketingReservationServiceMessage.java
@@ -1,4 +1,4 @@
-package com.example.mq.message;
+package com.example.publish.message;
 
 import java.util.List;
