+++ app/api/show-api/src/main/java/com/example/show/service/ShowAdminService.java
@@ -2,8 +2,8 @@ package com.example.show.service;
 
 
 import com.example.component.FileUploadComponent;
-import com.example.mq.MessagePublisher;
-import com.example.mq.message.ShowRelationArtistAndGenreServiceMessage;
+import com.example.publish.MessagePublisher;
+import com.example.publish.message.ShowRelationArtistAndGenreServiceMessage;
 import com.example.show.error.ShowError;
 import com.example.show.service.dto.request.ShowCreateServiceRequest;
 import com.example.show.service.dto.request.ShowUpdateServiceRequest;
