+++ app/api/show-api/src/test/java/artist/service/ArtistServiceTest.java
@@ -13,8 +13,8 @@ import artist.fixture.dto.ArtistResponseDtoFixture;
 import com.example.artist.service.ArtistService;
 import com.example.artist.service.dto.request.ArtistSubscriptionServiceRequest;
 import com.example.artist.service.dto.request.ArtistUnsubscriptionServiceRequest;
-import com.example.mq.MessagePublisher;
-import com.example.mq.message.ArtistSubscriptionServiceMessage;
+import com.example.publish.MessagePublisher;
+import com.example.publish.message.ArtistSubscriptionServiceMessage;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.UUID;
