+++ app/api/show-api/src/test/java/genre/service/GenreServiceTest.java
@@ -11,8 +11,8 @@ import static org.mockito.Mockito.verify;
 import com.example.genre.service.GenreService;
 import com.example.genre.service.dto.request.GenreSubscriptionServiceRequest;
 import com.example.genre.service.dto.request.GenreUnsubscriptionServiceRequest;
-import com.example.mq.MessagePublisher;
-import com.example.mq.message.GenreSubscriptionServiceMessage;
+import com.example.publish.MessagePublisher;
+import com.example.publish.message.GenreSubscriptionServiceMessage;
 import genre.fixture.GenreRequestDtoFixture;
 import genre.fixture.GenreResponseDtoFixture;
 import java.util.List;
