+++ app/api/show-api/src/test/java/show/service/ShowAdminServiceTest.java
@@ -10,8 +10,8 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
 import com.example.component.FileUploadComponent;
-import com.example.mq.MessagePublisher;
-import com.example.mq.message.ShowRelationArtistAndGenreServiceMessage;
+import com.example.publish.MessagePublisher;
+import com.example.publish.message.ShowRelationArtistAndGenreServiceMessage;
 import com.example.show.service.ShowAdminService;
 import com.example.show.service.dto.request.ShowCreateServiceRequest;
 import com.example.show.service.dto.request.ShowUpdateServiceRequest;
