+++ app/infrastructure/build.gradle
@@ -10,5 +10,5 @@ allprojects {
 dependencies {
     implementation project(":app:infrastructure:redis")
     implementation project(":app:infrastructure:s3")
-    implementation project(":app:infrastructure:message_queue")
+    implementation project(":app:infrastructure:message-queue")
 }
