+++ app/infrastructure/message-queue/src/main/java/org/example/config/PubSubConfig.java
