+++ app/infrastructure/message-queue/src/main/java/org/example/message/ShowRelationArtistAndGenreInfraMessage.java
@@ -1,6 +1,6 @@
 package org.example.message;
 
-import com.example.mq.message.ShowRelationArtistAndGenreServiceMessage;
+import com.example.publish.message.ShowRelationArtistAndGenreServiceMessage;
 import java.util.List;
 import java.util.UUID;
 import lombok.Builder;
