+++ settings.gradle
@@ -17,4 +17,4 @@ include(":app:api:show-api")
 include (":app:infrastructure")
 include (":app:infrastructure:redis")
 include (":app:infrastructure:s3")
-include (":app:infrastructure:message_queue")
+include (":app:infrastructure:message-queue")
